- Consolidated a few headers files. - Removed old SFML headers. - Moved window creation code out of the engine. - Fixed examples to compile again.

+++ inc/vapor/render/SFML_Window.h
@@ -78,7 +78,7 @@ private:
 	sf::Window window;
 	sf::WindowSettings windowSettings;
 	sf::VideoMode videoMode;
-	ulong flags;
+	uint32 flags;
 	
 	// Holds the current state of the cursor.
 	bool cursorState;
